+++ CHANGES.rst
@@ -9,7 +9,7 @@ Be sure to upgrade to 1.x before upgrading to 2.x.
 * Wazimap now supports multiple releases of a dataset. This means that you can, for example, have two releases of a national census in Wazimap at the same time. You will need to make some changes to how you build profile pages.
 * BREAKING: SimpleTable and FieldTable are now Django models.
 * BREAKING: Table management is now done through the Django admin interface. You must upgrade your existing tables.
-* BREAKING: ``get_objects_by_geo`` has been replaced by `FieldTable.get_objects_by_geo`.
+* BREAKING: ``get_objects_by_geo`` has been replaced by ``FieldTable.get_rows_for_geo``.
 
 1.1.1 (21 November 2017)
 -----------------------